requirements.txt
# 核心依赖
openai>=1.35.0
json-repair>=0.25.0
python-docx>=1.1.0
tiktoken>=0.7.0
jsonschema>=4.25.1
PyYAML>=6.0.0

# 可选依赖（用于某些功能）
requests>=2.31.0
click>=8.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0